refactor: 일반/상세 검색결과 domain 모델 네이밍 수정
