+++ app/src/main/java/com/teamwss/websoso/domain/mapper/NovelMapper.kt
@@ -2,11 +2,11 @@ package com.teamwss.websoso.domain.mapper
 
 import com.teamwss.websoso.data.model.ExploreResultEntity
 import com.teamwss.websoso.data.model.ExploreResultEntity.NovelEntity
-import com.teamwss.websoso.domain.model.NormalExploreResult
-import com.teamwss.websoso.domain.model.NormalExploreResult.Novel
+import com.teamwss.websoso.domain.model.ExploreResult
+import com.teamwss.websoso.domain.model.ExploreResult.Novel
 
-fun ExploreResultEntity.toDomain(): NormalExploreResult {
-    return NormalExploreResult(
+fun ExploreResultEntity.toDomain(): ExploreResult {
+    return ExploreResult(
         resultCount = resultCount,
         isLoadable = isLoadable,
         novels = novels.map { it.toDomain() },
